server/policy.py
"""
Policy Engine for Burly MCP Server

This module implements the security policy system that controls which tools
can be executed and validates their arguments. The policy engine loads
configuration from YAML files and enforces security constraints.

Key Components:
- PolicyLoader: Reads and parses policy/tools.yaml
- SchemaValidator: Validates tool arguments against JSON schemas
- ToolDefinition: Data model for tool configuration
- Security enforcement for path traversal, timeouts, and output limits

The policy system is the core security mechanism that prevents unauthorized
operations and ensures all tool executions are properly validated and logged.

Example policy.yaml structure:
```yaml
tools:
  docker_ps:
    description: "List Docker containers"
    args_schema:
      type: "object"
      properties: {}
    command: ["docker", "ps", "--format", "table"]
    mutates: false
    requires_confirm: false
    timeout_sec: 30
    notify: ["success", "failure"]
```
"""

from dataclasses import dataclass
from typing import Dict, List, Any, Optional
import yaml
import jsonschema
from pathlib import Path


@dataclass
class ToolDefinition:
    """
    Configuration for a single tool that can be executed via MCP.
    
    This dataclass defines all the security and operational parameters
    for a tool, including its command, validation schema, and execution
    constraints.
    """
    name: str
    description: str
    args_schema: Dict[str, Any]
    command: List[str]
    mutates: bool
    requires_confirm: bool
    timeout_sec: int
    notify: List[str]


class PolicyEngine:
    """
    Main policy engine that loads tool definitions and enforces security.
    
    The policy engine is responsible for:
    - Loading tool configurations from YAML files
    - Validating tool arguments against JSON schemas
    - Enforcing security constraints and limits
    - Providing tool definitions to the execution engine
    """
    
    def __init__(self, policy_path: str = "policy/tools.yaml"):
        """
        Initialize the policy engine with the specified policy file.
        
        Args:
            policy_path: Path to the YAML policy configuration file
        """
        self.policy_path = policy_path
        self.tools: Dict[str, ToolDefinition] = {}
        
        # TODO: Implement policy loading
        # self._load_policies()
    
    def _load_policies(self) -> None:
        """
        Load tool policies from the YAML configuration file.
        
        This method reads the policy file, validates its structure,
        and creates ToolDefinition objects for each configured tool.
        """
        # TODO: Implement YAML loading and validation
        pass
    
    def validate_tool_args(self, tool_name: str, args: Dict[str, Any]) -> bool:
        """
        Validate tool arguments against the tool's JSON schema.
        
        Args:
            tool_name: Name of the tool to validate
            args: Arguments to validate
            
        Returns:
            True if validation passes, False otherwise
        """
        # TODO: Implement JSON schema validation
        return True
    
    def get_tool_definition(self, tool_name: str) -> Optional[ToolDefinition]:
        """
        Get the tool definition for the specified tool name.
        
        Args:
            tool_name: Name of the tool to retrieve
            
        Returns:
            ToolDefinition if found, None otherwise
        """
        # TODO: Implement tool lookup
        return None
    
    def list_available_tools(self) -> List[str]:
        """
        Get a list of all available tool names.
        
        Returns:
            List of tool names that can be executed
        """
        # TODO: Return list of configured tools
        return []