pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "burly-mcp"
version = "0.1.0"
description = "A secure, policy-driven Model Context Protocol (MCP) server for system operations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Burly MCP Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.20.0",
    "requests>=2.31.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/burly-mcp"
Repository = "https://github.com/your-org/burly-mcp"
Issues = "https://github.com/your-org/burly-mcp/issues"

[project.scripts]
burly-mcp = "server.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=server",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]